requirements.txt
Django==5.0.4
python-environ==0.4.54
psycopg2==2.9.9
pillow==10.3.0
djangorestframework==3.15.1